+++ src/services/cart/ICartService.kt
@@ -6,7 +6,7 @@ import com.gmail.marcosav2010.model.SessionCart
 
 interface ICartService {
 
-    fun updateProductAmount(session: Session, productId: Long, amount: Long, add: Boolean): SessionCart
+    fun updateProductAmount(session: Session, productId: Long, amount: Int, add: Boolean): SessionCart
 
     fun remove(session: Session, productId: Long): SessionCart
